Retry s3.exists checks up to 3 times
s3.exists will sometimes return false negatives with young files. The
hypothesis is that s3 hasn't quite succeeded in the atomicity it wishes
for and is being eventually consistent. Under this hypothesis, we
suspect that we'd get fewer false positives if we try a couple of times
rather than just once. Problem is, this will be slower.

Add an option to s3.exists to enable repeat checking. Make it default to
on for now. For the moment, use the same connection, as creating a new
connection each time will be slower. We'll have to see if the problem
goes away. If it doesn't, then we'll open a new ticket to try it with a
new connection each time.

Merge of 63_s3_exists_retry / PR 2

Squashed commit of the following:

commit d0b23663cb08ed18562d3fd347df4ae5a5e84ad7
Author: Alex Weiss <algrs@cacography.net>
Date:   Wed Mar 23 15:11:41 2016 -0400

    Bump version

commit 82b1b64b34c1fb7d3e8b75388a826ea7f356823d
Author: Alex Weiss <algrs@cacography.net>
Date:   Wed Mar 23 15:10:08 2016 -0400

    Updates from PR

commit 7603408104773073fb9fbdf689b7d3634b48d185
Author: Alex Weiss <algrs@cacography.net>
Date:   Tue Mar 22 15:53:17 2016 -0400

    Updates from PR

    Add some info on how much this slows things down.

    Also add a --retries option to the s3 command and speed things up a hair
    by removing the bucket existance check on the retries, saving a HEAD for
    each subsequent attempt.

commit ce7fcdc698255c889fde95def50c490cd277dad6
Author: Alex Weiss <algrs@cacography.net>
Date:   Tue Mar 22 14:17:01 2016 -0400

    Add a warning in case we want to papertrail this

commit e6a1f2a662917e38e7f38efea261103f53e8dbd1
Author: Alex Weiss <algrs@cacography.net>
Date:   Tue Mar 22 13:49:50 2016 -0400

    Retry s3.exists checks up to 3 times